all-in-one-script/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.26.0
beautifulsoup4==4.10.0
pytube==11.0.1
PyGithub==1.55
httpx==0.23.0

